First draft for a viewer for the mods.de forum
Add the basic structure for a viewer for the mods.de forum. The idea
is to use the XML API to obtain the required information from the
server, and fill a couple of mustache templates that are displayed in
a WebView. So far only the part for reading has been implemented:
- the overview over all boards
- the list of threads in one board
- the list of posts in one thread

Rudimentary elements for navigation have been implemented:
- swipe left/right in board and thread view to change the page
- reload via the floating action button
- go back in history via the floating action button
This makes heavy use of the jQuery and jQuery mobile libraries.

The input to the WebView is provided via a local HTTP server,
unfortunately webview_flutter does not allow access to local files
(even if a page is set via a data: uri).

The BB code of a post is parsed and transformed into HTML. Smileys
and thread icons are available locally to save on data volume.
Caching for Benders has been implemented, they will only be
downloaded in case the device is connected via WiFi.