Require Flutter version 1.1.5
The previous stable version of Flutter had an issue on iOS: if an
application is paused and later resumed, widgets appear to not update
on the screen, while their internal state is updated. In particular
this affects the app bar title displaying the current forum or thread
name which is stuck at the last title before the pause. Require the
current stable version 1.2.1 of Flutter which seems to be the first
stablle version containing a fix for this (Flutter issue 24900).

As a consequence, require at least version 2.1.0 of dart, using the
previous requirement caused a warning about 'Future' only being
available in a later version (the version of dart in Flutter 1.0.0
still was a developer version).

Use this opportunity to set a particular range of versions for
dependencies.

Reformat the complete code.