Start HTTP server when app is started
Make sure the HTTP server is started when the application is started.
In the previous version content was only shown after the app was
paused and later resumed on Android. On iOS the content was always(?)
shown. Apparently, when an iOS app is started from memory the
application state is 'resumed'. Starting the HTTP server in this case
is covered by observing that state. Loading an app onto the phone for
debugging apparently leads to the app being in memory, so the cold
start case never happened in my tests.