Add 'toBBCode' functionality to BB code parser
Add 'toBBCode' functions to BBCodeTag and BBCodePart classes. This
can be used to obtain a BB code representation of the current
document. It is not necessarily equivalent to the original code, but
is the "corrected" (or whatever the parser thinks is correct) version
of it. This change very clearly show that the argument handling needs
to be improved.